+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.8] - 2024-09-17
 ### Fixed
 - Fixed `util.get_implemented` bug that was happening with the new version of `plum`
 
@@ -67,7 +69,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Renamed ode variable update, removed wrong test in utils
 - Updated support/packaging files
 
-[Unreleased]: https://github.com/ucl-bug/jaxdf/compare/0.2.7...master
+[Unreleased]: https://github.com/ucl-bug/jaxdf/compare/0.2.8...master
+[0.2.8]: https://github.com/ucl-bug/jaxdf/compare/0.2.7...0.2.8
 [0.2.7]: https://github.com/ucl-bug/jaxdf/compare/0.2.6...0.2.7
 [0.2.6]: https://github.com/ucl-bug/jaxdf/compare/0.2.5...0.2.6
 [0.2.5]: https://github.com/ucl-bug/jaxdf/tree/0.2.5
+